requirements.txt
pandas
SQLAlchemy
python-dotenv
psycopg2-binary